Move path following behavior to a separate module
