+++ src/movement.rs
@@ -4,6 +4,7 @@ use bevy::prelude::*;
 use crate::collisions::plugin::ColliderComponent;
 
 pub mod behaviors;
+pub mod paths;
 pub mod steering;
 
 #[derive(Component, Debug, Default, Clone, Copy, PartialEq, Reflect)]
